updates to add tests, add npm
update readme
