+++ README.md
@@ -1,3 +1,15 @@
 # flattenArray
 
-This method flattens a multidimensional array.
+This method extends the `Array.prototype` with the purpose of flattening a multidimensional array.
+
+
+### How to run?
+```sh
+npm install
+```
+
+### How to test?
+```sh
+npm run test
+```
+>>>>>>> updates to add tests, add npm
