+++ index.js
@@ -1,0 +1,2 @@
+'use strict';
+require('/src/flatten');
